Code cleaned, now oximeter is no more hardcoded
Note that there is a problem when getting the bioData body. If you get the reference of bioData, the access will return you values that does not make any sense. Therefore, I get the bioData OBJECT, NOT REFERECE, from the oximeter, and then I access the bioData object with the normal dot notation.